Refactored my profile page so its not 2.5k lines anymore. Hopefully this makes cursor better at inserting it as context.
