+++ lib/widgets/common/sliver_tab_bar_delegate.dart
@@ -1,0 +1,3 @@
+class SliverTabBarDelegate extends SliverPersistentHeaderDelegate {
+  // Move _SliverAppBarDelegate implementation here
+}
