+++ lib/widgets/profile/user_list_item_skeleton.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/material.dart';
+
+class UserListItemSkeleton extends StatelessWidget {
+  const UserListItemSkeleton({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return ListTile(
+      leading: const CircleAvatar(
+        backgroundColor: Colors.grey,
+      ),
+      title: Container(
+        width: 100,
+        height: 16,
+        decoration: BoxDecoration(
+          color: Colors.grey[300],
+          borderRadius: BorderRadius.circular(4),
+        ),
+      ),
+      subtitle: Container(
+        width: 150,
+        height: 12,
+        margin: const EdgeInsets.only(top: 4),
+        decoration: BoxDecoration(
+          color: Colors.grey[200],
+          borderRadius: BorderRadius.circular(4),
+        ),
+      ),
+    );
+  }
+}
