+++ lib/widgets/profile_tabs/videos_grid.dart
@@ -1,0 +1,1 @@
+// Extract video grid logic and UI
