+++ pubspec.yaml
@@ -53,7 +53,7 @@ dependencies:
   rxdart: ^0.27.7
   video_compress: ^3.1.2
   image: ^4.1.3
-  share_plus: ^7.0.0
+  share_plus: ^7.2.1
   intl: ^0.18.0
   chewie: ^1.7.5
   image_cropper: ^5.0.1
